backend/pyproject.toml
[project]
name = "focus-sprint-coach"
version = "0.1.0"
description = "Focus Sprint Coach backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "httpx>=0.27.0",
  "jinja2>=3.1.0",
  "python-multipart>=0.0.9",
  "sqlalchemy[asyncio]>=2.0.0",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt<4.1.0",
  "python-jose[cryptography]>=3.3.0",
  "email-validator>=2.1.0",
  "sentry-sdk>=2.0.0",
  "stripe>=8.0.0",
  "duckduckgo-search>=7.0.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
  "black>=24.0.0",
  "mypy>=1.8.0",
  "types-python-dateutil>=2.8.19.14",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
plugins = []
warn_return_any = true
warn_unused_configs = true
no_implicit_optional = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["app/tests"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*"]
